lab: changed gpmdown file and added refactoring sources to read

+++ util/gpmdown.py
@@ -1,29 +1,30 @@
 from typing import Sequence
 import requests
 from multiprocessing.pool import ThreadPool
-
 from requests.sessions import Session, session
+from os.path import exists
 
 username = "adisen99"
 password = "Sai@baba99"
 
 def download_url(url):
-    print("downloading: ", url)
     file_name_start_pos = url.rfind("/") + 1
-    file_name_end_pos = url.rfind("?") - 1
+    file_name_end_pos = url.rfind("?")
     filename = url[file_name_start_pos:file_name_end_pos]
 
-    with requests.Session() as session:
-        session.auth = (username, password)
-        r1 = session.request('get', url)
-        r = session.get(r1.url, auth=(username, password))
-        if r.ok:
-            with open(filename, 'wb') as f:
-                for data in r:
-                    f.write(data)
+    if not exists(filename):
+        print("downloading: ", url)
+        with requests.Session() as session:
+            session.auth = (username, password)
+            r1 = session.request('get', url)
+            r = session.get(r1.url, auth=(username, password))
+            if r.ok:
+                with open(filename, 'wb') as f:
+                    for data in r:
+                        f.write(data)
 
 # Get the urls from the txt file as a list using readline
-file1 = open('myfile.txt', 'r')
+file1 = open('links.txt', 'r')
 Lines = file1.readlines()
 
 urls = []
@@ -34,6 +35,6 @@ for line in Lines:
     urls.append(line.strip())
 
 # Run 5 multiple threads. Each call will take the next element in urls list
-results = ThreadPool(5).imap_unordered(download_url, urls)
+results = ThreadPool(6).imap_unordered(download_url, urls)
 for r in results:
-    print(r)
+    print("Completed" + r)
